requirements.txt
# ============================================
# LangGraph Voice Agents - Dependencies
# NO OpenAI - Using Cartesia for Voice
# ============================================

# Core Framework
langgraph>=0.2.0
langchain>=0.2.0
langchain-core>=0.2.0

# Voice Processing - Cartesia (NO OpenAI)
cartesia>=1.0.0
pyaudio>=0.2.14
livekit>=0.11.0
numpy>=1.24.0
websockets>=12.0

# LLM Providers (NO OpenAI)
anthropic>=0.25.0
langchain-anthropic>=0.2.0
google-generativeai>=0.5.0
langchain-google-genai>=1.0.0
httpx>=0.27.0  # For OpenRouter API

# State Management and Persistence
sqlalchemy>=2.0.0
redis>=5.0.0

# Data Processing
pydantic>=2.0.0
python-dotenv>=1.0.0
typing-extensions>=4.8.0

# Async and Concurrency
aiofiles>=23.0.0
aiohttp>=3.9.0

# Development Tools
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.0.0
ruff>=0.1.0
mypy>=1.5.0

# Optional: Enhanced Audio Processing
# librosa>=0.10.0
# soundfile>=0.12.0

# Optional: Monitoring
# prometheus-client>=0.17.0
# structlog>=23.0.0
